move miner init to ceremony_util
