+++ miner/src/lib.rs
@@ -26,5 +26,6 @@ pub mod submit_tx;
 pub mod test_tx_swarm;
 pub mod backlog;
 pub mod node_keys;
+pub mod keygen;
 // pub mod submit_noop_helper;
